backend/backend/requirements.txt
fastapi==0.68.2
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.8.2
python-multipart==0.0.5
email-validator==1.1.3
firebase-admin
numpy
scipy
librosa
opencv-python
bcrypt==3.2.0
